tests/wpt/tests/html/canvas/element/reset/2d.reset.state.shadow_color.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.reset.state.shadow_color</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.reset.state.shadow_color</h1>
<p class="desc">check that the state is reset</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("check that the state is reset");
_addTest(function(canvas, ctx) {

  const default_value = ctx.shadowColor;

  ctx.shadowColor = '#ff0000';
  _assert(ctx.shadowColor == '#ff0000', "ctx.shadowColor == '#ff0000'");

  ctx.reset();
  _assert(ctx.shadowColor == default_value, "ctx.shadowColor == default_value");

});
</script>
